travel_planning_system/car_rental_agent_langgraph/requirements.txt
langgraph==0.2.0
langchain-core==0.2.0
langchain-google-genai>=2.0.0
python-dotenv
requests
fastapi
uvicorn
pydantic
groq
langchain-groq>=0.3.0
langchain>=0.2.0
langchain-community>=0.2.0 